backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
psycopg==3.1.9
psycopg-binary==3.1.9
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
